+++ Plutarch/String.hs
@@ -19,12 +19,18 @@ import GHC.Generics (Generic)
 import Plutarch.Bool (PBool, PEq, (#&&), (#<=), (#==), (#||))
 import Plutarch.ByteString (PByteString)
 import Plutarch.Integer (PInteger)
-import Plutarch.Internal (S, Term, phoistAcyclic, (#), (:-->))
 import Plutarch.Internal.Lift (DeriveBuiltinPLiftable, PLiftable, PLifted (PLifted), pconstant)
 import Plutarch.Internal.Newtype (PlutusTypeNewtype)
 import Plutarch.Internal.Other (POpaque)
 import Plutarch.Internal.PLam (plam)
 import Plutarch.Internal.PlutusType (DPTStrat, DerivePlutusType, PlutusType)
+import Plutarch.Internal.Term (
+  S,
+  Term,
+  phoistAcyclic,
+  (#),
+  (:-->),
+ )
 import Plutarch.Unsafe (punsafeBuiltin)
 import PlutusCore qualified as PLC